+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - value count for each tab
 - visual marker for timezone-dependant values
 - "What's New" under Settings (references CHANGELOG.md on GitHub main branch, = this file)
-- option to expand the date value type with calendar week and day of the year
+- option to expand the date value type with ISO calendar week and day of the year
 
 ### Changed
 - up tp 4 custom tabs rather than 3